This is the first step of merging glimmer-cism-lanl/trunk into glimmer-cism2/branches/glimmer-cism-lanl.  This branch will ultimately be frozen and released.
Note: In many locations there were conflicts between the to branches.  I have tried to make the appropriate choices in these cases.  I tried to add comments as much as possible.  Grep for !EIB! to find places I modified/added/removed.  

This branch currently compiles.  However, this is not a statement on the correctness of the output.  Please go through the code carefully (especially where I commented) to verify the correctness.


git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/glimmer-cism-lanl@402 2cfc7cc4-626f-0410-8325-f6cb4dace9bf