Added changes to support build of glimmer with Trilinos on Jaguar.
For linking a code with Fortran main() with C++ linkers, an
additional library is needed. While this was picked up 
automatically on GCC, on Jaguar I had to manually add
in:  /opt/pgi/10.3.0/linux86-64/10.3/lib/f90main.o

I am committing in "jaguar-config" script that encodes
my build on Jaguar, since it there are a few magic pieces.




git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/glimmer-cism-lanl@1010 2cfc7cc4-626f-0410-8325-f6cb4dace9bf