Rebuilt Makefile.in files from Bluefire.
Bootstrapped on Bluefire to rebuild all of the Makefile.in files.  Could not bootstrap and build these successfully on Jaguar even with the latest Autotools.

This version has a new configure flag called --enable-mpi.  It is either set to "yes" or "no" to indicate whether parallel_mpi.F90 or parallel_single.F90 is used.  The parallel_*.F90 files are in libglimmer folder.  

This has been tested and found to configure and build with both values of --enable-mpi on Bluefire.  I had to make additions to the bf-config script in configure-scripts that I don't like.  I added NetCDF linking info to my configure variables in order to get it to configure.  I tried all of the versions on Bluefire, but finally settled on 4.1.2.
 
The type of configure error I got otherwise was
	checking for f90 netCDF interface... no
	configure: error: cannot find netCDF fortran library

Jeff



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1132 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 402b9881ffd7d0f2f5e744855267f63dd2e7d5ba