Got NOX working in SEACISM branch. Merged in improvements from merge-2.0-ncar branch into this branch.
Removed support in the libglimmer-trilinos directory for the 
returnGlobalVector option, which was needed for the old approach
of having a redundant-serial glimmer. Removed old capability
of translating the serial Triad matrix into Trilinos Epetra
format. Put in some better error checking and handling for 
input xml file.  Added nonlinear solver section to
  trilinosOptions.xml
file, both in source code and dome directories. This file
replaces   strat1.xml  and input.xml.

Need to acknowledge Valgrind tool which pointed me to
the last memory bug.

(Any experts on how we bootstrap these days? I need to 
remove a file but was afraid to touch Makefile.am.)




git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1167 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 3d08b6918f355248366030710c595e7f89e21284