Added nonlinear solver section to the trilinosOptions.xml for the confined shelf. Now this runs for JFNK.
Adding these lines to the xml file fixes the issue Steve saw with:

Throw number = 1
>> Error - The "Test Type" is a required parameter in the
>> NOX::StatusTest::Factory!


So, make sure the "Piro" section of trilinosOptions.xml is
filled in to use JFNK. If this section is empty, then the
input file is only set up for linear solves and not nonlinear.
Sorry the error was not caught at a reasonable level.




git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1195 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 88dd2e84682aebef1fcac9c93968ad440e83abf2