Computing coordinate vectors and passing to Trilinos.
In glam_strs2.F90 and glam_strs2.F90.osbc, I allocated
and computed  myX, myY, myZ arrays alongside myIndices.
These are the (scaled) values of the coordinates for
the nodes. For large problems, these become critical
for speedup with ML. For ML to make full use of these
coordinates, we will require Zoltan to be compiled into 
Trilinos, so I added that to the Trilinos config scripts.

Unless I messed up, this change does not effect
anyone right now. Irina and I will learn to use
these coordinates with ML, get the settings right,
and let you know when and how to use them.




git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1227 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 85d080cc20091d9cd54abb247e09d85d13259d45