Added Continuation in stress regularization paramter using LOCA.
On the Trilinos side, added the ability to do continuation
in "Effstrmin Factor" parameter. I set reasonable defaults so
it can be toggled with just one line in trilinosOptions.xml 

    <Parameter name="CISM: Number of Time Steps To Use LOCA" type="int" value="0"/>

In glam_strs2.F90, added an accessor subroutine to modify the
Effstrminsq value by the continuation parameter esm_factor
by this formula:

  effstrminsq = effstrminsq_target * 10.0**(2.0 * esm_factor)

When esm_factor=0.0 we have reached the standard settings 
for NOX solves. Defaults are to continue this factor from
10 to 0.0 with an initial step of -3.0.

Works great for dome case. On to GIS later.




git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1257 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 681709f010d8e2ef774d00915ab3f1f8f5a1aeb6