Made some minor changes that allow the new glissade_transport driver (including a parallel-friendly remapping scheme) to be used for thickness and temperature transport on a single processor with MPI/Trilinos. The code is now ready for parallelizing.
Running the dome test case, I verified for a range of cases 
(including temperature = 1 and 3, flow_law = 0 and 2, and 
which_ho_nonlinear = 0 and 1) that the new remapping scheme
gives the same results as the old scheme, within roundoff,
when the settings are identical. The results differ slightly
when I use a more accurate method to compute departure points
in the new scheme. 

Also, I tested the first-order upwind option and verified that 
it is working.  I have not yet done a formal comparison between 
upwind and remapping.

To use the new scheme, simply set old_remapping = .false. in glam.F90.
I left old_remapping = .true. in the version I am checking in now,
so users should not notice any differences.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1259 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: b4f2ca40daec557ff8e9a9c59cf241e413013186