Changed continuation parameter for LOCA runs to be an augmented viscosity.
FOr the continuation to be robust, we want the problem
to be easy for large values of the parameter, and become
the desired problem as the parameter reaches 0.0. 
The previous parameter  effstrmin, had the problem that
for large initial values it corresponded to very small
viscosities and presumably large velocities. By continuing
in an augmented viscosity, the initial solitions are for
large constant viscosities, which I think should be easier
to solve for.

I was able to converge the 5km_sliding problem for the
first time. For gis_2km, I still can not converge down 
to the real conditions.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1276 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 7decc225f2cf828d6c3f082ae5d3718e8aece3b1