Changed screen output line for partitioning information in inittrilinos. Prints out Total Max Min Ave  Nodes per processor.
The following is in progress and should not change anything:
Added flag to calc_F rountine to signal if it is a residual
calculation or a perturbed calculation for a directional
derivative. I am trying to base the preconditioner always
on the unperturbed residual instead of on the most recent
call. Initial attempts show identical behavior, but I think
this must be lead to an improvement eventually. 



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1278 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 842e2713c49d4bd85d0ef6164bcd2b33cc3ab313