Code commits related to scaling parameters.  Please read if you are a seacism developer. These commits are answer-changing and require a fresh bootstrap/config.
I am working toward simplifying and eventually removing the model scaling, which has
long been a source of frustration and bugs.

In glimmer_paramets.F90, I set thk0 = len0 = 1, with vel0 = 1/scyr.
These are the values that previously were set using the NO_RESCALE ifdef option;
now they are the default values.  So the computed model thickness (model%geometry%thck)
now has units of meters, and the computed velocity (model%velocity%uvel/vvel)
has units of meters/year.

Model results are different with the new scaling parameters, but only at the roundoff level.
After 10 timesteps with the dome test case, velocities are the same to ~12 digits
for the SIA solver and ~8 digits for the serial HO solver.

Previously there were two viscosity scaling paramters, vis0 and vis0_glam.
Now there is a single parameter called vis0 = tau0**(-gn) * (vel0/len0),
which was the old value of vis0_glam.

All scaling parameters can now be written in terms of scyr and (in the case of tau0)
rhoi*grav.  Without much additional work, we could remove these parameters 
(thk0, len0, vel0, acc0, tim0, vis0, evs0) from the code.  
I haven't removed them yet, because I'd like to do some more testing first.
We will have to decide if we want all SI units in the code, or if we want
to keep velocity in m/yr instead of m/s.  The former probably is less error-prone.

In several places, the previous code was hardwired to work only with the old 
Glimmer values of the scaling parameters.  I fixed these cases so that the code 
now works with either the old or the new values.  The old values are decremented
but can be restored by uncommenting a few lines in glimmer_paramets.F90.

The serial JFNK solver has trouble converging on the first timestep of the dome
problem with vel0 = 1/scyr.  (But it does eventually converge.)  With smaller
values of vel0, it may not converge.  But this may not be an issue if we decide
to remove the serial JFNK solver.

A related change: In glimmer_scales, I changed the names of the scaling parameters
that are used to scale I/O, and I removed some unused parameters.  
Previously, these had names like scale2d_f1.  The new names include a paradigm 
field with that particular scaling, e.g., scale_uvel.

A few other modules have cosmetic changes unrelated to scaling.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1373 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 6dd1da7ea8aefd4aefab83323e911d4e77c17737