Made merge of Bill's cleaned up glam_strs2.F90 and older .osbc1 version the new default version of glam_strs2.F90. The un-merged version is archived as glam_strs2.F90.old. Significant changes are marked with comments following the string "OSBC" (grepping for "OSBC" should reveal most changes).
Most of the changes to glam_strs2.F90 are in the subroutine "bodyset", which calls all of the internal and
boundary discretization subroutines and the subroutines that ulimately place coeffs. in the LHS matrix and 
soure terms into the RHS source vector.

Most of the boundary condition discretization subroutines have been altered to (1) allow for one-sided, 2nd-order
accurate finite diffs. at sfc and basal boundaries (so in the vert. dimension only at this point) and (2) to explicitly
include the effective visc. (efvs) term in the LHS matrix coeffs. These were prev. divided through so that they appeared
in the denom. of coeffs. attached to the RHS source terms. Conceptually, keeping them on the LHS should mean that
coeffs. associated with the sfc and basal bcs are now more similar to coeffs. for points internal to the ice sheet domain, 
since those coeffs. all contain an efvs. term.

Some subroutines that are no longer called were moved to the bottom of the module, into a section where it is noted that
the subroutines are not needed in this version of the code but stored for archive purposes.

Additional comments were added with a "TODO" flag, where chunks of code might be removed in the near future. This mainly 
has to do with logic surrounding subroutine calls for extracting blocks of off diagonal matrix coeffs. for
preconditioning of JFNK. This work was started but never fully debugged/finished and so has never been implemented.

With the old scaling in place, the code was tested for the dome and confined shelf test cases and gave nearly identical 
performance to test runs w/ slightly older versions of the .osbc1 versions of the code.

Also removed the version of glam_strs2.F90.osbc1 that has been stored on the repo, since those changes are now in 
glam_strs2.F90 (i.e., the code will now be built with the one-sided diff sfc and basal bcs by default).

The following "mixed" versions of the code are also still archived:

glam_strs2.F90.osbc1_homotopy
glam_strs2.F90.osbc2
glam_strs2.F90.osbc2_homotopy

The versions with the "osbc2" extension contain additional changes necessary to include one-sided basal bcs at the sfc 
and bed when at a lateral boundary (which also use one-side diffs in the horizontal). This was working at some point but
has been flakey for newer builds of the code so I suggest NOT using for now. I will try to test further and, if
performance improves, integrate w/ default version of glam_strs2.F90 at some point in the future.

I think that the 'homotopy' changes are (mostly) self contained and could be integrated with the new version of
glam_strs2.F90 with a small amount of work. 



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1382 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: d42a7e90b7fac94297e5a1332e5d4f88db6619ed