Merging seacism_timestep_reordering into the main seacism branch.
***NOTE: This commit will change answers in the seacism branch.***


This commit introduces substantial changes to glissade.F90, glide.F90 and simple_glide.F90 and some minor changes to other modules.  The purpose of this commit is to put various fields at the proper time level in the output.  For the most part it does not change calculations but some fields will be written at a different time level than before.

This resolves two separate issues related to time-stepping:
1. Diagnostic solve for initial time.  Previously output at time 0 was of a state advanced in time from the state passed in by the input file.  Now there is a diagnostic solve (of velocity and other diagnostic fields like usrf, lsrf) prior to time-stepping.  For a run where tend=tstart, the time-stepping loop is not entered at all.  This was accommodated by changing the looping condition and by adding diagnostic solve subroutines for both glissade and glide that are called in simple_glide prior to time-stepping.  

2. Within a time step there were inconsistencies in the time level in the output between the various variables in glissade.  Specifically, thickness was advanced one dt in time relative to velocity and temperature in the output.  This has been fixed by reordering when various calculations are made.  At the highest level, the new order is: 1. evolve column temperature, 2. evolve thickness/tracer advection, 3. diagnostic solve of velocity (and other stuff).

I have tested both glissade and glide to confirm that answers change only in the expected way due to the reordering.  With constant temperature or no temperature evolution, this means the same values just occur at a different time level of the output for some variables (for e.g. thickness, velocity).  With a changing temperature fields, there are some small changes to the answers.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1468 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 8a68ba07111e4fabf6250fafdfa2173b29b77ae7