Removed the umfpack and pardiso options from glimmer_sparse.F90. These were used for the old PBJ dycore but will not be supported for the release.
In their place, I added two new SLAP solver options:
* PCG with a diagonal preconditioner
* PCG with an incomplete Cholesky preconditioner

These solvers can be used to solve Ax = b when A is symmetric positive-definite,
as will be the case for the Glissade solver under construction.

I verified that the two new PCG solvers give a correct solution for simple 2x2 and 3x3
symmetric test matrices.  However, I had to set itol = 1 instead of itol = 2 to avoid
an error message.  For the PCG solvers, itol = 1 will be the default.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1495 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 9402bdc88301cbcfec77c387fc03577245b38e90