Checking in my latest work on the glissade finite-element solver.
For a dome test problem with zero sliding and constant viscosity, 
I now have what seems to be a correct velocity solution from the
linear SLAP solver.  I will work next on the outer nonlinear loop.

For testing constant viscosity, I added a new option 2 for whichefvs.
The existing option 1 makes the effective viscosity proportional to
the flow factor, so that it is not truly constant.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1531 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 8004b8aa4dc6f62b0094ba078c2ad2ea291bc747