Added a new module, cism_sparse_pcg.F90. This is a mostly home-grown PCG solver that is independent of SLAP. It doesn't currently add new code functionality (it uses diagonal preconditioning that doesn't work well for the dome problem), but it gives me a sandbox for trying some other things.  The new solver is called by setting which_ho_sparse = 5 in the config file.
In glimmer_sparse.F90 and glimmer_sparse_slap.F90, I changed 'matrix'
from intent(inout) to intent(in).  This avoids returning the matrix
with a modified data structure, with resulting errors in the residual
calculation. Instead, a local copy is made of the arrays passed to SLAP.

I also added some optional diagnostics that are turned off by default.
  


git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1575 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: a41ed93160b15e1182355629c3b6e3840a48e0e2