Added functionality to output 3d fields of the residual when using JFNK. Fields that can be output are those for the u,v vels (ures, vres) and the resid. magnitude (magres). By default, all of this code is commented out, including the variable declarations, allocations, and deallocations (in glide_types). Eventually, these bits of code could be surrounded by ifdef statements to allow for turning them on/off at compile time.
Other minor changes to glam_strs2 include slight altering of prev. commit which preformed a check on the value of
betasquared input from the .nc file, to make sure it was zero at areas of floating ice and g.l. The latter constraint
has been removed so for consistency with shelf test cases.

In glide_vars.def, simply added code so that the 3d res fields can be written to .nc output if desired.

Changes to res_vect were purely cosmetic.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1579 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 822ff08ab40350f86b68cb5be9970524600b78aa