Latest changes in glissade_velo_higher module and associated solvers.
In glide_types and glide_setup I added a new option, which_ho_approx, that specifies
whether the glissade solver is to solve the SIA, SSA, or HO equations.  In principle
it should work for all three (though it won't be the most efficient SIA or SSA solver).
I've tested it for HO and SIA but not yet for SSA.

The cism_sparse_pcg module includes a new subroutine, pcg_solver_structured,
that takes as input a set of matrices and vectors that live on the structured
(i/j/k) mesh.  (That is, they have not been converted to triad format.)
The linear matrix problem Ax = b is solved using a PCG solver based on the POP
barotropic solver.
  
The default diagonal preconditioner does not work well for the dome problem, 
so I added a physics-based preconditioner using the local shallow-ice part 
of the matrix.  The new preconditioner works well for the dome problem. 
At some point we could try some fancier physics-based preconditioners.




git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1586 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 0ebd11d7cb20512b46973be952bd61bd6f30c9ab