These commits are mostly related to code parallelism.
In glissade_transport.F90, I added parallel halo updates that are needed
for the global conservation checks to work correctly.

I added a test_parallel subroutine in glissade.F90.
This subroutine tests the ability of the code to advect a block of ice in an
arbitrary direction until it reaches the boundary and (with cyclic BCs)
reenters at the other side of the domain.

With the default parallel_halo updates, I verified that the IR advection works correctly
with cyclic (i.e., doubly periodic) boundary conditions on 1, 2, and 4 processors.

I parallelized the write_diagnostics subroutine in glide_diagnostics.F90.
This routine now correctly computes the total ice area, volume, etc.
on multiple processors.
Given a global diagnostic point (idiag, jdiag) in the config file,
the new code determines which local processor contains that point and
communicates the local ice state from that processor to the main processor,
which writes out the diagnostic info.

In glide_types.F90 and glide_setup.F90, I renamed idiag and jdiag to idiag_global
and jdiag_global, to make clear that the values in the config file
should be global (not local) horizontal coordinates.  At initialization,
the code checks that these coordinates are in bounds.

I removed the hardwired idiag and jdiag in glimmer_paramets.F90.
In simple_glide.F90 and in glint_main.F90, the arguments passed to write_diagnostics
are now model%numerics%idiag_global and model%numerics%jdiag_global.

(In simple_glide.F90, I also commented out the calls to glimmer_horiz_bcs, at least
for now.  After testing, I will also remove the parallel_halo calls if I can
show they are not needed.)

In parallel_mpi.F90, I added a flexible interface for parallel_reduce_max,
which now can work with integers and real4 as well as real8.
I added a similar interface in parallel_slap.F90.

In both parallel_mpi and parallel_slap, I added a function
called parallel_global_ID_scalar.  This is similar to the existing
function parallel_globalID, except that it assigns an ID of zero
to non-physical cells in the global halo.  Later we may
want to remove one of these two functions.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1600 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 52e8296733ee3d42ddbf3648574d2d714bba3460