Add ability to use default i/o files rather than running python script, needed for CESM build. I have added some default i/o files by copying the files that were autogenerated by the python script in a recent build.
Also, reworked how the FORTRAN_AUTOGEN_SOURCES are added to the
FORTRANSOURCES list in the cmake build, in a way that allows these to
be overridden by CESM SourceMods. The new method involves explicitly
specifying all of the sources that could possibly end up in
FORTRAN_AUTOGEN_SOURCES. While this is not very maintainable, we
should be able to remove it once people are no longer using
directories in which they have done autotools-based builds or older
cmake-based builds. (It would be possible to change this logic so that
it is more dynamic, similar to how SOURCEMODs are handled, but that
would involve adding some ugly logic for something that is only meant
as a short-term workaround.)

Tested 4 builds on hopper (all cmake builds):
- gnu from a clean checkout
- pgi from a clean checkout
- gnu from a directory that had old autogenerated files in-source
- pgi from a directory that had old autogenerated files in-source


git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1658 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: c3fd5d59bf1f3cd1c32a4fdf5cceecf6a0c697b6