Fix ocean check in glint_i_tstep: this needs to do the proper bit-wise check of the mask, rather than checking for exact equality.
With this fix, acab & artm are identical between (a) a CESM/CISM run
with the old codebase (pre-cism2), and (b) a CESM/CISM run with the
seacism codebase, using cism1 (SIA) config options. (Caveat: these are
only identical if thickness evolution is off: with thickness evolution
on, they differ, presumably because of divergent thickness evolution
in the two codebases.)

There are still substantial differences in prognostic fields between
the two codebases, and these differences need to be resolved.

I have tested this change in CESM, but not with a standalone run.


git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1662 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 3813c134893775056cac591cffde803821f399d7