This check-in includes some year-end progress on the new glissade finite-element solver (whichdycore = 2).  We can now solve a dome problem in parallel using MPI without Trilinos. I've tested the parallel solver for up to 4 processors on my Mac, and the results agree within roundoff to those obtained with 1 processor.
These are the sparse solver options now supported:
  integer, parameter :: HO_SPARSE_BICG = 0
  integer, parameter :: HO_SPARSE_GMRES = 1
  integer, parameter :: HO_SPARSE_PCG_INCH = 2
  integer, parameter :: HO_SPARSE_PCG_STRUC = 3
  integer, parameter :: HO_SPARSE_TRILINOS = 4

The new solver option is HO_SPARSE_PCG_STRUC = 3.  
(Numbers 2 and 3 were left available after we removed support for Pardiso
and Umfpack.  Number 2 is a recently added SLAP option using PCG with
incomplete Cholesky preconditioning.)

So far I've tested the new parallel solver only for a standard dome problem, 
but in principle it should work for any problem of the form Ax = b, provided that
the matrix A is symmetric positive-definite.

The new solver subroutines are in libglimmer-solve/cism_sparse_pcg.F90.
The driver subroutine is based on the preconditioned conjugate gradient solver
in the POP ocean model.	It solves the problem Ax = b, where A, x, and b
use the (k,i,j) structure of the Cartesian grid	(hence 'STRUC' in
the option name).  Since all arrays are	structured, we can use the
existing halo update routines for the global matrix-vector multiplies
and dot products required by PCG.

The default preconditioner is a shallow-ice solver that operates on a
simplified local-column version of the matrix A, with horizontal coupling
removed.  It works reasonably well (unlike diagonal preconditioning,
which does not work well for the dome problem).

This new solver is not very general.  It contains only a PCG solver and
thus requires an SPD matrix, and it likely will not scale well to large problems.  
But it could be useful for code development and debugging.

In glimmer_sparse.F90 and glimmer_sparse_slap.F90 I removed the PCG_DIAG
option I added a few weeks ago, because PCG with diagonal preconditioning
doesn't work well compared to PCG with incomplete Cholesky.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1664 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 8f8ffe2d3e44f400552ec52b2e49bd1d3cd7d3a9