Changes to allow exact restarts in the GLIDE dycore (dycore=0).
Exact restarts will now occur for the evolution=0 (linear diffusion) and evolution=2 (nonlinear diffusion) evolution schemes.  They are NOT exact for evolution=1 (ADI).  For evolution=0, uvel and vvel must be included as restart variables to ensure exact restart.  A warning is issued saying so if that option is chosen.  Eventually this could be handled by the proposed new implementation of creating the hotvars list.

This commit will changes answers relative to older versions of the code when using evolution=2, but the changes are within the convergence tolerance of the Picard iteration in that scheme.

Also added a fatal error if one tries to run GLIDE with IR evolution.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1677 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 0f1e453f6621b1ccf7f8faade19641baaf0cbde4