Implemented a new procedure for assigning restart variables.  Substantial changes were made to the way output files are setup, but users should not notice.
In the past, the necessary variables for a restart were set by indicating hot=1 in glide_vars.def.  However in the current code with multiple dycores and myriad physics options, maintaining a single master list of necessary restart variables has become confusing.  In that system, the list of necessary restart variables was determined at compile time, but I am here replacing it with a system that creates that list at run time, with the goal of having smaller, less confusing restart files.  This should also be easier to maintain than the old system.

With this commit, all 'hot' entries in glide_vars.def have been eliminated, along with the machinery to build the list of hotvars in generate_ncvars.py.  This has been replaced by subroutines in glimmer_ncparams.F90 that 1) analyze the config options at runtime and build an appropriate list of restart variables, and 2) determine if an output file is meant to be a restart file, and if so, includes the appropriate list of restart variables and sets the xtype to double.  I have also replaced the terminology of hotstarts and hotvars with restart and restart variables to use preferred language and emphasize the changes made.  

Going forward, a user needs to merely include the word 'restart' or 'hot' in the variable list of an output file, and the code will automatically setup that output file as a restart file.  Note that it is still ok to include additional variables in the variable list, but including either one of these words will trigger the file to become a restart file (which simply means the things listed in point 2 above will happen).  Going forward, the preferred terminology is 'restart', but the ability to use the word 'hot' has been retained for backward compatibility.  Here is an example of a config that would provide an output file that will allow you to do exact restarts:
[CF output]
variables = restart
frequency = 1 
name = dome.out.nc

A final note: define_restart_variables() in glimmer_ncparams.F90 is where the logic lives about which variables need to be restart variables.  I have tested some basic config options for exact restart (e.g. the range of dycore & evolution options), but some further analysis/testing may be necessary for some config combinations.  Please make any changes here as needed, and comment the reasons why a particular variable needs to be a restart variable.

Also I have not touched the functionality of options%hotstart which determines if the run being started is a restart/hotstart (as opposed to if an output file is meant to be used for a future restart).

Lastly, a make clean is necessary.





git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1691 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 1287a76ef0fb18b434b0f21ea3dfaf7e91b43095