Fixing another problem with exact restarts in GLIDE.
After further review I determined that both wvel and wgrd should be restart variables with the glide dycore.  This is because the temp driver calculates weff=f(wvel,wgrd).  To simplify things, I have commented out the calls to the wgrd & wvel calculations in the temp driver and placed them entirely at the end of glide tstep.  It may be possible to structure the calls so that wvel is not needed as a restart variable (see notes in glimmer_ncparams.F90), but I think it is less confusing to implement it the way I am committing.  define_restart_variables() has been modified to automatically make both wvel and wgrd restart variables if the GLIDE dycore is chosen.

I also adjusted the initialization of temperature and flwa in glide_temp to follow the same structure as I implemented in glissade_temp in a recent commit.

I have confirmed exact restarts of temperature (and thk and velo) for all combinations of evolution=0,2 and vertical_integration=0,1.

This commit requires a make clean because of a change to glide_vars.def.

The change to the location of the wvel calculations causes small changes to temperature calculations relative to older versions of the code, which propagate to changes in thk and vel after the first time step.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1693 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 0d32fea9478926618fd17cf691d373569cde5828