Making it allowable to run the GLIDE dycore (and associated temperature option) in a parallel build if only one processor is used.  This is handy for, among other things, switching between both dycores without having to recompile (assuming one wants to run glissade in parallel).
I've tested this against the output of the serial build and the output matches with one exception.  In thkmask the serial version gives a value of 132 (land+domain edge) in a one cell outline around the edge of the domain.  The parallel version just gives 4 (land) for those cells.  In other words, the parallel version does not identify the edge of the domain in thkmask.  

I also confirmed that the parallel version (1 proc) performs exact restarts for GLIDE.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1694 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 9e7af8801a1d60a04a38e3f699c0293b3bd98ddc