Making further modifications to the way restart variables are handled to make the system compatiable with CESM glc code.
Expansion of the restart variable list has been returned to ncdf_template.F90.in (which is the template for *_io.F90) so that the expansion does not happen until a file is created.  This was important for glc code that creates restart files during timestepping.

Determination of the restart variable list has been moved from glimmer_ncparams.F90 to glide_setup.F90/isostasy_setup.F90 and happens right after the config options have been read.  Note that now the restart variable list itself has been moved to *_io.F90, and each *_io.F90 module has its own unique list that gets filled out during config.  I still need to extend this functionality to glint - so far it just has been applied to glide and glide_lithot.  Additions to each module's restart variable list are made through a call to *_add_to_restart_variable_list() which are located in *_io.F90.

I have also added checks in glimmer_config.F90 and *_io.F90 for config option strings and variable lists that are too long for the variables that hold them.  This should alleviate problems of truncated variable lists (or at least let the user know they won't get all the variables they asked through a fatal error).  I also increased the length that a config line can be, but we may find we want to increase it further.  (Note the length that can be read and the length of variables that hold config options are different and both may have to be adjusted - this should probably be made consistent at some point.)



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1701 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 02764d013769d4acf003120e2497c2a5fe3919a0