Added a fatal error message in case someone tries to run the elastic lithosphere with multiple processors (tasks > 1).
It would take a lot of work to parallelize the elastic lithosphere or put
something better in its place.  This is on the back burner for now.

A local column lithosphere model is also available (by setting lithosphere = 0).
This might not be hard to parallelize, and I'll look at it later.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1706 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 6bb03ae48b8fd8c175af7aba5eb5a2034c49d9a9