Changes to what happens on a restart (when restart=1 in the config file).
The most important is that output is not written at the initial time.  This means there will not be a duplicate time level in the two output files before and after a restart.

Also changes to how some variables are initialized in certain situations (temp, velocity, flwa) so that restarts are performed correctly.  This has been tested for common config options but not for all physics options (e.g. not for basal water or marine_margin options).

Also changed terminology within the code - 'hotstart' is no longer used anywhere and has been replaced with 'restart'.  Note that hotstart=1 (instead of restart=1) has been retained for backward compatability only.  (So has the variable name 'hot' instead of 'restart'.)



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1721 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 971491ec5fa4f114142fb0613f6b9260f2bf5935