Set of code changes to allow periodic lateral boundary conditions when using parallel code with seacism dycore.
Changes are as follows:

- ALL calls to "horiz_bcs ..." subroutines have been left in the code but are currently commented OUT. This is becuase
  1) in most cases they did not appear to be doing anything and 2) existing code works to implement parallel lateral
bcs. Further work will need to be done to understand if these subroutine calls are needed for additional lateral bc
functionality and/or if they can recreate the periodicity that currently works when they are not called. Subroutines for
which this was the ONLY change relative to prev. versions of the code include glide_velo.F90, glide_derivs.F90, 
glam_strs2.F90, and glissade_velo.F90.

- in glide_mask.F90, parallel logic to mark domain boundaries as such was commented out. This was resulting in some
  lateral boundary points being marked as "land margin", in which case the wrong logic was being triggered in the
assembly subroutines (i.e., this is why "valueset" was being triggered). I think this bit of code can probably be
removed at this point. It looks like it was mainly in there to allow for similar results between the serial and parallel
codes in cases where the whole domain was filled with ice, which we intend to deal with differently now.

- in glissade.F90, a set of halo updates was applied to the thck and usrf fields on the NON-staggered grid. This was
  immediately followed by the calculation of 1) the staggered thickness field and 2) thickness and usrf gradients on
that staggered mesh, that use non-staggered usrf and thck as inputs. This was the main bug fix that allowed for the
periodic bcs to be implemented correctly; without this, the stag thickness array and (possibly) the usrf and thck
gradient arrays on the stag mesh had incorrect values in their global halos. These changes were made in the
glissade_diagnostic_variable_solve(model) subroutine call and are extensively commmented. In addition, halo updates to
the fields lsrf and topg were also added, followed by calls to the subroutines that calculate staggered values
of these same vars (e.g, calculating staglsrf from lsrf). This was done for consistency, because spot checking indicated
that these stag fields also did not have the correct information in their global halos. It doesn't appear that they are
currently needed to allow the periodic bc implementation to work. Additional comments were added to exisiting comments 
regarding if/where halo updates should be applied.  
 
Bill L. may want to confirm that is dycore still exists with these updates, since his dycore also uses glissade.F90 and
glissade_velo.F90. 

This commit will be followed by an udpate for the ismip-hom test case .py and .config scripts.




git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1726 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 54dda7265078f25e1d286dd0289f8d2d96b3066d